ggml-qnn: rebase to upstream
